[client,sdl] move prefs to static lib
This is in preparation for the subsequent commit adding test case for
the prefs functionality to avoid building the code twice.
