+++ client/SDL/CMakeLists.txt
@@ -63,10 +63,10 @@ find_package(SDL2 REQUIRED COMPONENTS)
 include_directories(${SDL2_INCLUDE_DIR})
 include_directories(${SDL2_INCLUDE_DIRS})
 
-set(LIBS "")
+set(PREFS_LIBS "")
 if (cJSON_FOUND)
     include_directories(${CJSON_INCLUDE_DIRS})
-    list(APPEND LIBS ${CJSON_LIBRARIES})
+    list(APPEND PREFS_LIBS ${CJSON_LIBRARIES})
     add_compile_definitions(CJSON_FOUND)
 endif()
 
@@ -95,8 +95,13 @@ set(SRCS
     sdl_window.cpp
 )
 
+add_library(sdl_prefs STATIC
+    sdl_prefs.hpp
+    sdl_prefs.cpp
+)
+
 add_subdirectory(aad)
-list(APPEND LIBS
+set(LIBS
         winpr
         freerdp
         freerdp-client
@@ -104,6 +109,7 @@ list(APPEND LIBS
         sdl_client_res
         dialogs
         aad-view
+        sdl_prefs
     )
 
 if (NOT WITH_SDL_LINK_SHARED)
@@ -115,6 +121,7 @@ endif()
 AddTargetWithResourceFile(${PROJECT_NAME} "${WIN32_GUI_FLAG}" "${PROJECT_VERSION}" SRCS)
 
 target_link_libraries(${PROJECT_NAME} PRIVATE ${LIBS})
+target_link_libraries(sdl_prefs ${PREFS_LIBS})
 set_property(TARGET ${PROJECT_NAME} PROPERTY FOLDER "Client/SDL")
 install(TARGETS ${PROJECT_NAME} DESTINATION ${CMAKE_INSTALL_BINDIR} COMPONENT client)
